+++ packages/atomic/src/components/search/facets/atomic-popover/atomic-popopver.stories.tsx
@@ -1,0 +1,15 @@
+import {html} from 'lit-html';
+import defaultStory from '../../../../../.storybook/default-story';
+
+const {defaultModuleExport, exportedStory} = defaultStory(
+  'Atomic/Popover',
+  'atomic-popover',
+  {},
+  {
+    additionalChildMarkup: () =>
+      html` <atomic-facet label="File types" field="filetype"></atomic-facet>`,
+  }
+);
+
+export default defaultModuleExport;
+export const DefaultPopover = exportedStory;